added mappers, started to add validation on db level via annotations

+++ src/main/java/islab1/models/DTO/EventDTO.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 public class EventDTO {
     private long id;
-    private Integer creatorId;
+    private long creatorId;
     private String name;
     private Integer minAge;
     private EventType eventType;
